# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Reviewly",
    "description": "Keep track of your reviews",
    "version": "0.1.0"
  },
  "host": "localhost:3000",
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded",
    "text/plain"
  ],
  "produces": [
    "application/json",
    "text/plain",
    "text/html"
  ],
  "basePath": "/api",
  "paths": {
    "/v1/search/{term}": {
      "get": {
        "x-swagger-router-controller": "searches",
        "operationId": "searchCompany",
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "tags": ["search"],
        "description": "Search for a company by name",
        "parameters": [
          {
            "name": "term",
            "in": "path",
            "description": "Search term of company you are looking for",
            "required": true,
            "type": "string"
          },
          {
            "name": "radius",
            "in": "query",
            "description": "Search radius to look for",
            "required": false,
            "type": "number"
          },
          {
            "name": "loc",
            "in": "query",
            "description": "Coorindates you are looking for ex: 0,0",
            "required": false,
            "type": "string"
          },
          {
            "name": "access_token",
            "description": "Internal access token",
            "in": "query",
            "type": "string",
            "required": false
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/401Response"
            }
          },
          "200": {
            "description": "Successful request."
          }
        }
      }
    },
    "/v1/reviews/{rid}": {
      "get": {
        "x-swagger-router-controller": "reviews",
        "operationId": "readReview",
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "tags": ["reviews"],
        "description": "Read a specific review by id",
        "parameters": [
          {
            "name": "rid",
            "in": "path",
            "description": "Review ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "access_token",
            "description": "Internal access token",
            "in": "query",
            "type": "string",
            "required": false
          }
        ],
        "responses": {
          "401": {
            "description": "Error"
          },
          "200": {
            "description": "Successful request.",
            "schema": {
              "$ref": "#/definitions/Review"
            }
          }
        }
      }
    },
    "/v1/reviews": {
      "get": {
        "x-swagger-router-controller": "reviews",
        "operationId": "readAllReviews",
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "tags": ["reviews"],
        "description": "Read all reviews for a user",
        "parameters": [
          {
            "name": "access_token",
            "description": "Internal access token",
            "in": "query",
            "type": "string",
            "required": false
          }
        ],
        "responses": {
          "401": {
            "description": "Error"
          },
          "200": {
            "description": "Successful request.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Reviews"
              }
            }
          }
        }
      },
      "post": {
        "x-swagger-router-controller": "reviews",
        "operationId": "createReview",
        "tags": ["reviews"],
        "description": "Create a new review",
        "parameters": [
          {
            "name": "review",
            "in": "body",
            "description": "Review information",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Review"
            }
          },
          {
            "name": "access_token",
            "description": "Internal access token",
            "in": "query",
            "type": "string",
            "required": false
          }
        ],
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "responses": {
          "401": {
            "description": "Authentication Failed"
          },
          "200": {
            "description": "Successful request."
          }
        }
      }
    },
    "/v1/images": {
      "post": {
        "x-swagger-router-controller": "images",
        "operationId": "createImage",
        "tags": ["images"],
        "description": "Upload image",
        "consumes": [
          "multipart/form-data",
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "file",
            "in": "formData",
            "type": "file",
            "description": "Image information",
            "required": false
          },
          {
            "name": "access_token",
            "description": "Internal access token",
            "in": "query",
            "type": "string",
            "required": false
          }
        ],
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "responses": {
          "401": {
            "description": "Error"
          },
          "200": {
            "description": "Successful request."
          }
        }
      }
    },
    "/v1/images/{iid}": {
      "get": {
        "x-swagger-router-controller": "images",
        "operationId": "readImageURL",
        "tags": ["images"],
        "description": "Return URL of image given specific image identifier",
        "parameters": [
          {
            "name": "iid",
            "in": "path",
            "type": "string",
            "description": "Image Identifier"
          },
          {
            "name": "access_token",
            "description": "Internal access token",
            "in": "query",
            "type": "string",
            "required": false
          }
        ],
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request.",
            "schema": {
              "url": "string"
            }
          },
          "401": {
            "description": "Error"
          }
        }
      }
    },
    "/v1/users": {
      "post": {
        "x-swagger-router-controller": "users",
        "operationId": "registerUser",
        "tags": ["users"],
        "description": "Register a new user",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "The user information for registration",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserRegistration"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Error"
          },
          "200": {
            "description": "Successful request.",
            "schema": {
              "_id": "string"
            }
          }
        }
      }
    },
    "/v1/users/{username}": {
      "get": {
        "x-swagger-router-controller": "users",
        "operationId": "profile",
        "tags": ["users"],
        "description": "Read profile of single user by passing in a username",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Read the profile of a specific username",
            "required": true,
            "type": "string"
          },
          {
            "name": "access_token",
            "description": "Internal access token",
            "in": "query",
            "type": "string",
            "required": false
          }
        ],
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "responses": {
          "401": {
            "description": "Error"
          },
          "200": {
            "description": "Successful request.",
            "schema": {
              "$ref": "#/definitions/userResponse"
            }
          }
        }
      }
    },
    "/v1/auths/facebook": {
      "post": {
        "x-swagger-router-controller": "auths",
        "operationId": "facebook",
        "tags": ["auths"],
        "description": "Facebook authentication",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Facebook Token",
            "required": true,
            "schema": {
              "$ref": "#/definitions/FacebookToken"
            }
          }
        ],
        "security": [
          {
            "facebookAuth": []
          }
        ],
        "responses": {
          "401": {
            "description": "Error"
          },
          "200": {
            "description": "Successful request.",
            "schema": {
              "$ref": "#/definitions/InternalToken"
            }
          }
        }
      }
    },
    "/v1/auths/basic": {
      "post": {
        "x-swagger-router-controller": "auths",
        "operationId": "basic",
        "tags": ["auths"],
        "description": "Provide username and password to get auth token.",
        "parameters": [
          {
            "name": "credentials",
            "in": "body",
            "description": "The username and password",
            "required": false,
            "schema": {
              "$ref": "#/definitions/UserCredentials"
            }
          }
        ],
        "security": [
          {
            "localAuth": []
          }
        ],
        "responses": {
          "401": {
            "description": "Error"
          },
          "200": {
            "description": "Successful request.",
            "schema": {
              "$ref": "#/definitions/InternalToken"
            }
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    },
    "localAuth": {
      "type": "basic",
      "description": "Local authentication via body parameters"
    },
    "tokenAuth": {
      "type": "apiKey",
      "description": "Built in token auth",
      "name": "access_token",
      "in": "query"
    },
    "facebookAuth": {
      "type": "apiKey",
      "description": "Facebook token auth using passport-facebook-token",
      "name": "access_token",
      "in": "query"
    }
  },
  "definitions": {
    "basicAuth": {
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "apiKey": {
      "properties": {
        "access_token": {
          "type": "string"
        }
      }
    },
    "Reviews": {
      "properties": {
        "_id": {
          "type": "string",
          "example": "54f44c5d26352b1b9755b12f"
        },
        "description": {
          "type": "string",
          "example": "This is a review of starbucks coffee, it could be 144 characters long"
        },
        "datetime": {
          "type": "string",
          "format": "date-time",
          "example": "2015-03-04T07:36:56.224Z"
        },
        "company": {
          "type": "string",
          "example": "Starbucks Coffee"
        },
        "images": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ImageURL"
          }
        }
      },
      "required": ["_id"]
    },
    "ImageURL": {
      "properties": {
        "url": {
          "type": "string",
          "example": "https://www.google.com/images/srpr/logo11w.png"
        }
      }
    },
    "Review": {
      "properties": {
        "company": {
          "type": "string",
          "example": "Starbucks, Polk Street"
        },
        "description": {
          "type": "string",
          "example": "My coffee was too cold and too sweet"
        },
        "rating": {
          "type": "number",
          "example": 2
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string",
            "example": "54f44c5d26352b1b9755b12f"
          }
        },
        "datetime": {
          "type": "string",
          "format": "date-time",
          "example": "2015-03-04T07:36:56.224Z"
        },
        "location": {
          "type": "string",
          "example": "50.3 -120.5"
        }
      },
      "required": ["company", "description"]
    },
    "InternalToken": {
      "properties": {
        "token": {
          "type": "string"
        }
      }
    },
    "FacebookToken": {
      "properties": {
        "access_token": {
          "type": "string"
        }
      }
    },
    "userResponse": {
      "properties": {
        "username": {
          "description": "Username of the user",
          "type": "string"
        },
        "profile": {
          "description": "User Profile Information",
          "type": "object",
          "$ref": "#/definitions/userProfile"
        }
      },
      "example": {
        "measurements": {
          "waist": "32.2",
          "hips": "36.2",
          "thighs": "34.1",
          "calves": "12.9",
          "length": "32.0",
          "inseam": "28.3"
        },
        "status": {
          "received": "yes",
          "measured": "pending"
        }
      }
    },
    "userProfile": {
      "properties": {
        "reviews": {
          "description": "Number of reviews the user has made",
          "type": "number"
        },
        "points": {
          "description": "Number of points the user has based on reviews and images",
          "type": "number"
        },
        "rank": {
          "description": "The name rank of the user",
          "type": "string"
        }
      },
      "example": {
        "reviews": "12",
        "points": "14",
        "rank": "Newbie"
      }
    },
    "UserCredentials": {
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "UserRegistration": {
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Ok": {
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "required": ["message"]
    },
    "401Response": {
      "properties": {
      }
    },
    "400Response": {
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "required": ["message"]
    },
    "Error": {
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "required": ["message"]
    }
  }
}